Convert ThreadLocal-based pool to parameterized version
