--- src/main/java/com/fasterxml/jackson/core/util/JsonBufferRecyclerPools.java
@@ -1,15 +1,0 @@
-package com.fasterxml.jackson.core.util;
-
-import com.fasterxml.jackson.core.JsonFactory;
-
-/**
- * Set of {@link BufferRecyclerPool} implementations to be used by the default
- * JSON-backed {@link JsonFactory} for recycling {@link BufferRecycler}
- * containers.
- *
- * @since 2.16
- */
-public final class JsonBufferRecyclerPools
-{
-
-}
